+++ client/src/App.js
@@ -1,5 +1,5 @@
 import GlobalStyle from './styles/GlobalStyles';
-import Header from './components/Header';
+import Header from './components/header/Header.js';
 
 const App = () => {
   return (
